Removed the CanFrame struct
When trying to use this library it was a lot of hassle to translate between the CanFrame struct and some other CanFrame struct needed to interace the hw. It seems like the interface is better off beeing traits, and the CanFrame struct only lives inside the tests module. The TransportFrame trait needs to be revamped, and it should also be considered adding a CanFrame : TransportFrame trait which is simpler to implement