first attempt at creating a type that transfers buffer references
Even if this current implementation is okay, I don't think it should be
exposed to crate users in the future. It's very easy to get the
semantics incorrect.

I believe it should work with any async setup (at least if both tasks
are running in the same executor (or different executors and there's no
preemption). If this is true, I might do some higher-level
implementation on top of it, but for now, it stays as is.

Definitely will need to create an OwnedTransfer type, that uses alloc
types, so you can easily transfer between threads.